+++ package-metadata.json
@@ -1,0 +1,1 @@
+{"sublime_text": "*", "description": "A Sublime Text 2 Plugin to run SVN commands", "platforms": ["*"], "version": "2017.11.15.01.03.16", "dependencies": [], "url": "https://github.com/LordBrom/SVNTools"}
